+++ app/src/main/java/com/vm/shadowsocks/core/NatSession.java
@@ -2,7 +2,7 @@ package com.vm.shadowsocks.core;
 
 public class NatSession {
     public int RemoteIP;
-    public short RemotePort;
+    public int RemotePort;
     public String RemoteHost;
     public int BytesSent;
     public int PacketSent;
